requirements.txt
wordcloud
google-api-python-client
pyarrow
fastparquet
gcsfs
alpha_vantage
ratelimit
transformers
gpustat
diffusers
yfinance
sentence-transformers
autogluon